Port the Serverless REST API w/ DynamoDB example
This is an initial cut at the Serverless REST API w/ DynamoDB example,
rewritten in Pulumi.  Note the nice way that routes and the database
can just use code, rather than CloudFormation.

Of course, the @pulumi/aws-serverless API abstraction has yet to be
born, however my sincere hope is that we can get that done this week.

I initially wrote this in TypeScript with modern ES6-style modules,
which is quite nice, however, for a better appples-to-apples comparison,
we should lead in with the pure JavaScript and require-based port.

Until I can run this for real, I'm sure there are some closure
serialization issues in here, like the places I'm capturing the AWS
SDK imported on the outside of the function.  I'll keep this as-is
until I can actually get it up and running because this is what 99%
of the people out there will think to do and I want to feel the
experience first hand of whether this works as expected or not.