Bug fix: Allow false to be set in the config (#1028)
The old code used the logical OR operator so that if the value in the config was false, it would be set to true. This made it impossible to set it to false in the config. This commit changes that to the nullish coalescing operator so that it will only default to true if no value is set.