Bump dependencies of aws-go-s3-folder (#1395)
The aws-go-s3-folder example is used for our perf metrics. It currently depends on an older version of pulumi-aws that has a bug that shows an unnecessary diff that results in an empty update after an initial update taking longer than our alerting threshold. This change updates the example to use the latest version of pulumi-aws with the fix (and also updates the similar aws-go-s3-folder-component example).