Do not run benchmark prewarm in parallel with the benchmark
Looks like ProgramTest defaults to enabling t.Parallel(), and this defeats the purpose of prewarm
phase in provider benchmarks. The intent of prewarm was to do a no-op to make sure dependencies are
downloaded and avoid measuring that overhead as part of the benchmark. This was not the case though
because prewarm was running in parallel with the actual benchmark.

After the change, each individual benchmark first does unmeasured prewarm, and then does the
measurement. In the future we may want to measure prewarm as it is interesting for user experience
as well.