set RequireService to true for benchmarks
RequireService used to be the default for ProgramTest, in older
pulumi/pkg versions.  However when pkg got updated in
https://github.com/pulumi/examples/commit/3d889a90972a440406524a785b0152fe8a1fb00d#diff-8af729bc718ea271aecd1ceaae6d16e2119d7dfbe6a21b6e6f793b3b92fc65d1R10,
running against the service started to be hidden behind a flag for
better test isolation.

We do however want to run benchmarks against the service, so set
RequireService to true here to get that behaviour back.