set pipefail option when running tests
When running the tests in CI, we're piping the output of `go test`
into gotestfmt to get prettier output.  This works because `gotestfmt`
will return a non-zero exit code when there are test failures.

However when there are other failures, e.g. compilation of the tests
failed, `gotestfmt` will have a 0 exit code.  `go test` wouldn't, but
because we're piping the results through, the shell swallows that
unless we're setting the pipefail option.

Set up the pipefail option here, so tests aren't accidentally green
when compilation fails.