update Go version to 1.21.x
pulumi/pulumi generally supports the two latest Go releases. Since 1.22.x came out, we dropped support for 1.20.x, and indeed compilation of the latest pulumi/pulumi source is failing with the latest Go version because of the missing slices package.

Update CI to the latest supported Go version, so we are able to update pulumi/pulumi in this repo.